chore: Move having messages back into its own component
